requirements.txt
aniso8601==9.0.1
attrs==22.1.0
CacheControl==0.12.11
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cleo==2.0.1
click==8.1.3
crashtest==0.4.1
cryptography==37.0.4
distlib==0.3.6
dulwich==0.20.46
exceptiongroup==1.1.0
filelock==3.8.0
Flask==2.2.3
flask-restx==1.0.6
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
html5lib==1.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.2
jsonschema==4.16.0
keyring==23.9.3
lockfile==0.12.2
MarkupSafe==2.1.2
marshmallow==3.19.0
more-itertools==8.14.0
msgpack==1.0.4
packaging==22.0
pexpect==4.8.0
pipenv==2022.12.19
pkginfo==1.9.2
platformdirs==2.5.2
pluggy==1.0.0
poetry==1.3.2
poetry-core==1.4.0
poetry-plugin-export==1.2.0
ptyprocess==0.7.0
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.18.1
pytest==7.2.2
pytz==2022.7.1
rapidfuzz==2.13.7
requests==2.28.1
requests-toolbelt==0.9.1
SecretStorage==3.3.3
shellingham==1.5.0
SQLAlchemy==2.0.4
tomli==2.0.1
tomlkit==0.11.5
trove-classifiers==2022.12.22
typing_extensions==4.5.0
urllib3==1.26.12
virtualenv==20.17.1
virtualenv-clone==0.5.7
webencodings==0.5.1
Werkzeug==2.2.3
